Simulator.PlanCreator behaviour (distributed_simulator v0.1.0)

Module which should be used by exactly one module in every simulation. That module will be called PlanCreator module.

Using module have to implement function create_plan/5 which will be responsible for creating plans for every dynamic object in the simulation. Function should return a tuple {direction, plan} where plan is a one dimensional tensor with two elements: [action, consequence].

See Evacuation.PlanCreator in the examples directory for the exemplary usage.

Callbacks

create_plan(x_index, y_index, grid, objects_state, iterations)

Specs

create_plan(
  x_index :: Simulator.Types.index(),
  y_index :: Simulator.Types.index(),
  grid :: Nx.t(),
  objects_state :: Nx.t(),
  iterations :: Simulator.Types.index()
) :: {direction :: Nx.t() | integer(), plan :: Nx.t()}