Simulator.PlanResolver behaviour (distributed_simulator v0.1.0)

Module which should be used by exactly one module in every simulation. That module will be called PlanResolver module.

It has to implement three functions:

is_update_valid?/2 - which will be responsible for checking whether the action can be applied to the object;

apply_action/3 - which should return new object and state for the location of the action of the plan.

apply_consequence/3 - which should return new object and state for the old object location.

See Evacuation.PlanResolver in the examples directory for the exemplary usage.

Callbacks

apply_action(object, plan, old_state)

Specs

apply_action(object :: Nx.t(), plan :: Nx.t(), old_state :: Nx.t()) ::
  {new_object :: integer(), new_state :: Nx.t()}

apply_consequence(object, plan, old_state)

Specs

apply_consequence(object :: Nx.t(), plan :: Nx.t(), old_state :: Nx.t()) ::
  {new_object :: integer(), new_state :: Nx.t()}

is_update_valid?(action, object)

Specs

is_update_valid?(action :: Nx.t(), object :: Nx.t()) :: Nx.t()