Simulator.Printer (distributed_simulator v0.1.0)

Module responsible for saving visualization data and metrics to the files. Additionally contatins functions helpful in debugging.

Functions

clean_grid_iterations()

Delete all the contents of the directory with files for visualization.

create_metrics_directory(location)

Creates directory for metrics of the worker located in {x, y}.

create_visualization_directory(location)

Creates directory for visualization of the grid of the worker located in {x, y}.

print(grid, phase \\ nil)

Prints given grid.

print_3d_tensor(tensor, description \\ nil)

Prints 3D tensor in readable way. Useful during debugging.

print_accepted_plans(accepted_plans, description \\ nil)

Prints accepted plans in readable way. Useful during debugging.

print_objects(grid, description \\ nil)

Prints only the objects from the given grid.

print_objects_state(objects_state, description \\ nil)

Prints objects state in readable way. Useful during debugging.

write_to_file(state)

Writes grid as tensor to file. Firstly, it is converted to string.