Simulator.WorkerActor (distributed_simulator v0.1.0)

GenServer responsible for simulating one shard.

There are three phases of every iteration:

:remote_plans - plans are processed. Some of them are accepted, some discarded. Result of the processing is distributed among neighboring shards;

:remote_consequences - consequences derived from the accepted plans are applied to the grid. Additionally, signal update is calculated and distributed to the neighboring shards;

:remote_signal - signal is applied to the grid. Next iteration is started. If iteration's number does not exceed the maximum number of iterations set in the cofiguration, plans are created and distributed to the neighboring shards;

First iteration starts with creating plans, so it is in the middle of :remote_signal phase.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

start(list)

Specs

start(keyword(Nx.t())) :: GenServer.on_start()

Starts the WorkerActor.

TODO use some supervisor.

unstash_messages(state)