Simulator.WorkerActor.Plans (distributed_simulator v0.1.0)

Module contataining Worker's functions responsible for the plans.

Each plan is a tensor: [direction, action, consequence].

direction - a plan contains an action towards a specific neighboring cell. Direction indicates it.

action - what should be the state of the target cell (pointed by direction).

consequence - what should be in the current cell if the plan will be executed.

Example: a person wants to move up: [@dir_up, @person, @empty].

Functions

create_plans(iteration, grid, objects_state, create_plan)

Specs

create_plans(Simulator.Types.index(), Nx.t(), Nx.t(), (... -> any())) :: Nx.t()

Creates plans for every cell in the grid.

process_plans(grid, plans, objects_state, is_update_valid?, apply_action)

Specs

process_plans(Nx.t(), Nx.t(), Nx.t(), (... -> any()), (... -> any())) ::
  {Nx.t(), Nx.t(), Nx.t()}

The function decides which plans are accepted and update the grid by putting action in the proper cells. Consequences will be applied in the :remote_consequences phase.