Simulator.WorkerActor.Signal (distributed_simulator v0.1.0)

Module contataining Worker's functions responsible for the signal.

Functions

apply_signal_update(grid, signal_update, signal_factor)

Specs

apply_signal_update(Nx.t(), Nx.t(), (... -> any())) :: Nx.t()

Applies signal update.

Cuts out only signal (without object) from grid and signal_update, performs applying update and puts result back to the grid.

Applying update is making such an operation on every signal value {i, j, dir}: s[i][j][dir] = (s[i][j][dir] + S u[i][j][dir]) A * f(g[i][j][0]) where:

s - a signal grid (3D tensor cut out from grid);

u - passed signal update (3D tensor);

g - passed grid;

S - @signal_suppression_factor;

A - @signal_attenuation_factor;

f - signal_factor function - returned value depends on the contents of the cell.

calculate_signal_updates(grid, generate_signal)

Specs

calculate_signal_updates(Nx.t(), (... -> any())) :: Nx.t()

Calculates signal update for all cells.